[FIX] use fix-sized integers in sam/bam

+++ include/seqan/bam_io/bam_sam_conversion.h
@@ -128,7 +128,7 @@ void _appendTagsSamToBamOneTag(TTarget & target, TForwardIter & iter, CharString
                     ++nEntries;
 
             // Write out array length.
-            appendRawPod(target, (uint32_t)nEntries);
+            appendRawPod(target, nEntries);
 
             // Write out array values.
             size_t startPos = 1;
